Cisco APIC ML2 mechanism driver, part 2
    This set of changes introduces a mechanism driver for the
    Cisco APIC. This is the second and final part of a 2 part commit.
    Please see the blueprint for more information.

    The review is submitted in two parts:
    - Part 1 (Posted earlier, required for Part 2)
        o APIC REST Client
        o APIC data model and migration script
        o APIC configurations
    - Part 2 (this commit)
        o APIC mechanism driver
        o APIC manager

    Partially implements: blueprint ml2-cisco-apic-mechanism-driver

+++ neutron/plugins/ml2/drivers/cisco/apic/apic_model.py
@@ -172,5 +172,6 @@ class ApicDbModel(object):
         """Deletes the port profile for a node."""
         profile = self.session.query(PortProfile).filter_by(
             node_id=node_id).first()
-        self.session.delete(profile)
+        if profile:
+            self.session.delete(profile)
         self.session.flush()
